+++ examples/squeezencnn/.gitignore
@@ -1,0 +1,5 @@
+.gradle/
+.idea
+build/
+local.properties
+.externalNativeBuild
