--- examples/squeezencnn/assets/squeezenet_v1.1.bin
@@ -1,1 +1,0 @@
-../../squeezenet_v1.1.bin
