--- examples/squeezencnn/assets/synset_words.txt
@@ -1,1 +1,0 @@
-../../synset_words.txt
