--- examples/squeezencnn/jni/Application.mk
@@ -1,7 +1,0 @@
-
-# APP_STL := stlport_static
-APP_STL := gnustl_static
-# APP_ABI := armeabi armeabi-v7a
-APP_ABI := armeabi-v7a
-APP_PLATFORM := android-9
-NDK_TOOLCHAIN_VERSION := 4.9
