+++ examples/squeezencnn/local.properties
@@ -1,10 +1,12 @@
-# This file is automatically generated by Android Tools.
+## This file is automatically generated by Android Studio.
 # Do not modify this file -- YOUR CHANGES WILL BE ERASED!
 #
 # This file must *NOT* be checked into Version Control Systems,
 # as it contains information specific to your local configuration.
-
-# location of the SDK. This is only used by Ant
+#
+# Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-sdk.dir=/home/nihui/osd/android-sdk-linux
+#Mon Jul 24 17:59:47 CST 2017
+ndk.dir=/Users/lizhangqu/AndroidNDK/android-ndk-r14b
+sdk.dir=/Users/lizhangqu/AndroidSDK
