--- examples/squeezencnn/src/com/tencent/squeezencnn/MainActivity.java
@@ -1,189 +1,0 @@
-// Tencent is pleased to support the open source community by making ncnn available.
-//
-// Copyright (C) 2017 THL A29 Limited, a Tencent company. All rights reserved.
-//
-// Licensed under the BSD 3-Clause License (the "License"); you may not use this file except
-// in compliance with the License. You may obtain a copy of the License at
-//
-// https://opensource.org/licenses/BSD-3-Clause
-//
-// Unless required by applicable law or agreed to in writing, software distributed
-// under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
-// CONDITIONS OF ANY KIND, either express or implied. See the License for the
-// specific language governing permissions and limitations under the License.
-
-package com.tencent.squeezencnn;
-
-import android.app.Activity;
-import android.os.Bundle;
-
-import android.content.Context;
-import android.content.Intent;
-import android.database.Cursor;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-import android.net.Uri;
-import android.provider.MediaStore;
-import android.util.Log;
-import android.view.View;
-import android.widget.Button;
-import android.widget.ImageView;
-import android.widget.TextView;
-
-import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-
-import com.tencent.squeezencnn.SqueezeNcnn;
-
-public class MainActivity extends Activity
-{
-    private static final int SELECT_IMAGE = 1;
-
-    private TextView infoResult;
-    private ImageView imageView;
-    private Bitmap yourSelectedImage = null;
-
-    private SqueezeNcnn squeezencnn = new SqueezeNcnn();
-
-    /** Called when the activity is first created. */
-    @Override
-    public void onCreate(Bundle savedInstanceState)
-    {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.main);
-
-        try
-        {
-            initSqueezeNcnn();
-        }
-        catch (IOException e)
-        {
-            Log.e("MainActivity", "initSqueezeNcnn error");
-        }
-
-        infoResult = (TextView) findViewById(R.id.infoResult);
-        imageView = (ImageView) findViewById(R.id.imageView);
-
-        Button buttonImage = (Button) findViewById(R.id.buttonImage);
-        buttonImage.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View arg0) {
-                Intent i = new Intent(Intent.ACTION_PICK);
-                i.setType("image/*");
-                startActivityForResult(i, SELECT_IMAGE);
-            }
-        });
-
-        Button buttonDetect = (Button) findViewById(R.id.buttonDetect);
-        buttonDetect.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View arg0) {
-                if (yourSelectedImage == null)
-                    return;
-
-                String result = squeezencnn.Detect(yourSelectedImage);
-
-                if (result == null)
-                {
-                    infoResult.setText("detect failed");
-                }
-                else
-                {
-                    infoResult.setText(result);
-                }
-            }
-        });
-    }
-
-    private void initSqueezeNcnn() throws IOException
-    {
-        byte[] param = null;
-        byte[] bin = null;
-        byte[] words = null;
-
-        {
-            InputStream assetsInputStream = getAssets().open("squeezenet_v1.1.param.bin");
-            int available = assetsInputStream.available();
-            param = new byte[available];
-            int byteCode = assetsInputStream.read(param);
-            assetsInputStream.close();
-        }
-        {
-            InputStream assetsInputStream = getAssets().open("squeezenet_v1.1.bin");
-            int available = assetsInputStream.available();
-            bin = new byte[available];
-            int byteCode = assetsInputStream.read(bin);
-            assetsInputStream.close();
-        }
-        {
-            InputStream assetsInputStream = getAssets().open("synset_words.txt");
-            int available = assetsInputStream.available();
-            words = new byte[available];
-            int byteCode = assetsInputStream.read(words);
-            assetsInputStream.close();
-        }
-
-        squeezencnn.Init(param, bin, words);
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data)
-    {
-        super.onActivityResult(requestCode, resultCode, data);
-
-        if (resultCode == RESULT_OK && null != data) {
-            Uri selectedImage = data.getData();
-
-            try
-            {
-                if (requestCode == SELECT_IMAGE) {
-                    Bitmap bitmap = decodeUri(selectedImage);
-
-                    Bitmap rgba = bitmap.copy(Bitmap.Config.ARGB_8888, true);
-
-                    // resize to 227x227
-                    yourSelectedImage = Bitmap.createScaledBitmap(rgba, 227, 227, false);
-
-                    imageView.setImageBitmap(yourSelectedImage);
-                }
-            }
-            catch (FileNotFoundException e)
-            {
-                Log.e("MainActivity", "FileNotFoundException");
-                return;
-            }
-        }
-    }
-
-    private Bitmap decodeUri(Uri selectedImage) throws FileNotFoundException
-    {
-        // Decode image size
-        BitmapFactory.Options o = new BitmapFactory.Options();
-        o.inJustDecodeBounds = true;
-        BitmapFactory.decodeStream(getContentResolver().openInputStream(selectedImage), null, o);
-
-        // The new size we want to scale to
-        final int REQUIRED_SIZE = 400;
-
-        // Find the correct scale value. It should be the power of 2.
-        int width_tmp = o.outWidth, height_tmp = o.outHeight;
-        int scale = 1;
-        while (true) {
-            if (width_tmp / 2 < REQUIRED_SIZE
-               || height_tmp / 2 < REQUIRED_SIZE) {
-                break;
-            }
-            width_tmp /= 2;
-            height_tmp /= 2;
-            scale *= 2;
-        }
-
-        // Decode with inSampleSize
-        BitmapFactory.Options o2 = new BitmapFactory.Options();
-        o2.inSampleSize = scale;
-        return BitmapFactory.decodeStream(getContentResolver().openInputStream(selectedImage), null, o2);
-    }
-
-}
